// src/proto/astarte2_196.ts
export interface freeSpinList {
    /**
     * @generated from protobuf field: repeated serverProto.freeData free = 1
     */
    free: freeData[];
}
// ...
export interface GameInfoAck {
    /**
     * 钱包信息列表，每个元素是一个 WalletInfo 消息
     *
     * @generated from protobuf field: repeated serverProto.WalletInfo WalletInfo = 1
     */
    walletInfo: WalletInfo[];
    /**
     * 偏好轮次显示信息，是一个 PreferRoundShow 消息
     *
     * @generated from protobuf field: serverProto.PreferRoundShow Prefer = 2
     */
    prefer?: PreferRoundShow;
    /**
     * 免费旋转剩余次数
     *
     * @generated from protobuf field: int32 FreeSpinRemain = 3
     */
    freeSpinRemain: number;
    /**
     * 额外信息，以字节数组形式存储
     *
     * @generated from protobuf field: bytes extraInfo = 5
     */
    extraInfo: Uint8Array;
    /**
     * 游戏的理论回报率
     *
     * @generated from protobuf field: double rtp = 6
     */
    rtp: number;
    /**
     * 解锁大奖的投注金额列表
     *
     * @generated from protobuf field: repeated double jpUnlockBet = 7
     */
    jpUnlockBet: number[];
    /**
     * 商城信息，是一个 MallInfo 消息
     *
     * @generated from protobuf field: serverProto.MallInfo mall = 8
     */
    mall?: MallInfo;
    /**
     * API 类型
     *
     * @generated from protobuf field: int32 ApiType = 9
     */
    apiType: number;
    /**
     * 免费旋转的投注金额
     *
     * @generated from protobuf field: double FreeSpinBet = 10
     */
    freeSpinBet: number;
    /**
     * 游戏商城信息，是一个 GameMallInfo 消息
     *
     * @generated from protobuf field: serverProto.GameMallInfo gameMall = 12
     */
    gameMall?: GameMallInfo;
    /**
     * 最大赔率
     *
     * @generated from protobuf field: double MaxOdd = 13
     */
    maxOdd: number;
    /**
     * 免费旋转列表信息，是一个 freeSpinList 消息
     *
     * @generated from protobuf field: serverProto.freeSpinList freeSpin = 14
     */
    freeSpin?: freeSpinList;
    /**
     * 免费旋转的类型
     *
     * @generated from protobuf field: int32 freeSpinType = 15
     */
    freeSpinType: number;
    /**
     * 锦标赛信息，是一个 Tournament 消息
     *
     * @generated from protobuf field: serverProto.Tournament tournament = 16
     */
    tournament?: Tournament;
}
/**
 * 定义钱包信息的消息结构
 *
 * @generated from protobuf message serverProto.WalletInfo
 */
export interface WalletInfo {
    /**
     * 货币编号，使用 int32 类型存储
     *
     * @generated from protobuf field: int32 currencyNumber = 1
     */
    currencyNumber: number;
    /**
     * 货币名称，使用 string 类型存储
     *
     * @generated from protobuf field: string currencyName = 2
     */
    currencyName: string;
    /**
     * 货币符号，使用 string 类型存储
     *
     * @generated from protobuf field: string currencySymbol = 3
     */
    currencySymbol: string;
    /**
     * 钱包中的金币数量，使用 double 类型存储
     *
     * @generated from protobuf field: double coin = 4
     */
    coin: number;
    /**
     * 投注金额列表，每个投注金额使用 double 类型存储
     *
     * @generated from protobuf field: repeated double bet = 5
     */
    bet: number[];
    /**
     * 单位，使用 double 类型存储
     *
     * @generated from protobuf field: double unit = 6
     */
    unit: number;
    /**
     * 比例，使用 double 类型存储
     *
     * @generated from protobuf field: double ratio = 7
     */
    ratio: number;
    /**
     * 费率，使用 double 类型存储
     *
     * @generated from protobuf field: double rate = 8
     */
    rate: number;
    /**
     * 上限，使用 double 类型存储
     *
     * @generated from protobuf field: double upper = 9
     */
    upper: number;
    /**
     * 下限，使用 double 类型存储
     *
     * @generated from protobuf field: double lower = 10
     */
    lower: number;
    /**
     * 周期，使用 double 类型存储
     *
     * @generated from protobuf field: double cycle = 11
     */
    cycle: number;
    /**
     * 小数位数，使用 int32 类型存储
     *
     * @generated from protobuf field: int32 decimal = 12
     */
    decimal: number;
}
/**
 * 定义偏好轮次展示信息的消息结构
 *
 * @generated from protobuf message serverProto.PreferRoundShow
 */
export interface PreferRoundShow {
    /**
     * 基础轮次相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double BaseRound = 1
     */
    baseRound: number;
    /**
     * Sigma 轮次相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double SigmaRound = 2
     */
    sigmaRound: number;
    /**
     * 变异系数相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double CV = 3
     */
    cV: number;
    /**
     * 命中率相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double HR = 4
     */
    hR: number;
    /**
     * 最大连赢轮次相关数值，使用 double 类型存储
     *
     * @generated from protobuf field: double MGR = 5
     */
    mGR: number;
    /**
     * 范围，使用 int32 类型存储
     *
     * @generated from protobuf field: int32 Range = 6
     */
    range: number;
    /**
     * 展示标志，使用 int32 类型存储
     *
     * @generated from protobuf field: int32 Show = 7
     */
    show: number;
}
/**
 * 商城信息
 *
 * @generated from protobuf message serverProto.MallInfo
 */
export interface MallInfo {
    /**
     * 价格赔率，使用双精度浮点数存储
     *
     * @generated from protobuf field: double PriceOdd = 1
     */
    priceOdd: number;
    /**
     * 描述类型，使用 32 位有符号整数存储
     *
     * @generated from protobuf field: int32 DescType = 2
     */
    descType: number;
    /**
     * 最大投注金额，使用双精度浮点数存储
     *
     * @generated from protobuf field: double MaxBet = 3
     */
    maxBet: number;
    /**
     * 替代 ID，使用 32 位有符号整数存储
     *
     * @generated from protobuf field: int32 AlterID = 4
     */
    alterID: number;
    /**
     * 显示标志，使用 32 位有符号整数存储
     *
     * @generated from protobuf field: int32 Show = 5
     */
    show: number;
}
/**
 * 游戏商城信息
 *
 * @generated from protobuf message serverProto.GameMallInfo
 */
export interface GameMallInfo {
    /**
     * 最大投注金额，使用双精度浮点数存储
     *
     * @generated from protobuf field: double MaxBet = 1
     */
    maxBet: number;
    /**
     * 价格赔率列表，每个元素为双精度浮点数
     *
     * @generated from protobuf field: repeated double PriceOdd = 2
     */
    priceOdd: number[];
    /**
     * 替代 ID 列表，每个元素为 32 位有符号整数
     *
     * @generated from protobuf field: repeated int32 AlterID = 3
     */
    alterID: number[];
}
/**
 * 定义 Tournament 消息
 *
 * @generated from protobuf message serverProto.Tournament
 */
export interface Tournament {
    /**
     * @generated from protobuf field: int32 id = 1
     */
    id: number;
    /**
     * @generated from protobuf field: double balance = 2
     */
    balance: number;
    /**
     * @generated from protobuf field: double initBalance = 3
     */
    initBalance: number;
    /**
     * @generated from protobuf field: double MinBet = 4
     */
    minBet: number;
    /**
     * @generated from protobuf field: double MaxBet = 5
     */
    maxBet: number;
    /**
     * @generated from protobuf field: serverProto.goalType goal = 6
     */
    goal: goalType;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp start = 7
     */
    start?: Timestamp;
    /**
     * @generated from protobuf field: google.protobuf.Timestamp end = 8
     */
    end?: Timestamp;
    /**
     * @generated from protobuf field: bool buyBonusOn = 9
     */
    buyBonusOn: boolean;
    /**
     * @generated from protobuf field: bool extraBetOn = 10
     */
    extraBetOn: boolean;
    /**
     * @generated from protobuf field: serverProto.categoryType category = 11
     */
    category: categoryType;
}
// ...
export enum goalType {
    /**
     * @generated from protobuf enum value: NONE = 0;
     */
    NONE = 0,
    /**
     * @generated from protobuf enum value: ACCWIN = 1;
     */
    ACCWIN = 1,
    /**
     * @generated from protobuf enum value: MULT = 2;
     */
    MULT = 2
}
// ...
export const freeSpinList = new freeSpinList$Type();
// ...
export const freeData = new freeData$Type();
// ...
export const GameInfoAck = new GameInfoAck$Type();
// ...
export const WalletInfo = new WalletInfo$Type();
// @generated message type with reflection information, may provide speed optimized methods
class PreferRoundShow$Type extends MessageType<PreferRoundShow> {
    constructor() {
        super("serverProto.PreferRoundShow", [
            { no: 1, name: "BaseRound", kind: "scalar", jsonName: "BaseRound", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 2, name: "SigmaRound", kind: "scalar", jsonName: "SigmaRound", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 3, name: "CV", kind: "scalar", jsonName: "CV", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 4, name: "HR", kind: "scalar", jsonName: "HR", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 5, name: "MGR", kind: "scalar", jsonName: "MGR", T: 1 /*ScalarType.DOUBLE*/ },
            { no: 6, name: "Range", kind: "scalar", jsonName: "Range", T: 5 /*ScalarType.INT32*/ },
            { no: 7, name: "Show", kind: "scalar", jsonName: "Show", T: 5 /*ScalarType.INT32*/ }
        ]);
    }
    create(value?: PartialMessage<PreferRoundShow>): PreferRoundShow {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.baseRound = 0;
        message.sigmaRound = 0;
        message.cV = 0;
        message.hR = 0;
        message.mGR = 0;
        message.range = 0;
        message.show = 0;
        if (value !== undefined)
            reflectionMergePartial<PreferRoundShow>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: PreferRoundShow): PreferRoundShow {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* double BaseRound */ 1:
                    message.baseRound = reader.double();
                    break;
                case /* double SigmaRound */ 2:
                    message.sigmaRound = reader.double();
                    break;
                case /* double CV */ 3:
                    message.cV = reader.double();
                    break;
                case /* double HR */ 4:
                    message.hR = reader.double();
                    break;
                case /* double MGR */ 5:
                    message.mGR = reader.double();
                    break;
                case /* int32 Range */ 6:
                    message.range = reader.int32();
                    break;
                case /* int32 Show */ 7:
                    message.show = reader.int32();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: PreferRoundShow, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* double BaseRound = 1; */
        if (message.baseRound !== 0)
            writer.tag(1, WireType.Bit64).double(message.baseRound);
        /* double SigmaRound = 2; */
        if (message.sigmaRound !== 0)
            writer.tag(2, WireType.Bit64).double(message.sigmaRound);
        /* double CV = 3; */
        if (message.cV !== 0)
            writer.tag(3, WireType.Bit64).double(message.cV);
        /* double HR = 4; */
        if (message.hR !== 0)
            writer.tag(4, WireType.Bit64).double(message.hR);
        /* double MGR = 5; */
        if (message.mGR !== 0)
            writer.tag(5, WireType.Bit64).double(message.mGR);
        /* int32 Range = 6; */
        if (message.range !== 0)
            writer.tag(6, WireType.Varint).int32(message.range);
        /* int32 Show = 7; */
        if (message.show !== 0)
            writer.tag(7, WireType.Varint).int32(message.show);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message serverProto.PreferRoundShow
 */
export const PreferRoundShow = new PreferRoundShow$Type();
// ...
export const MallInfo = new MallInfo$Type();
// ...
export const GameMallInfo = new GameMallInfo$Type();
// ...
export const Tournament = new Tournament$Type();